# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  types.CreateUserRequest:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      user_name:
        type: string
    required:
    - date_of_birth
    - email
    - password
    - user_name
    type: object
  types.EditUserRequest:
    properties:
      date_of_birth:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  types.LoginRequest:
    properties:
      password:
        type: string
      user_name:
        type: string
    required:
    - password
    - user_name
    type: object
  types.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  types.UserDetailInfoResponse:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
info:
  contact:
    email: maxuanquang@gmail.com
    name: Quang Ma
    url: https://www.linkedin.com/in/maxuanquang
  description: A simple social network management service API in Go using Gin framework
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://maxuanquang.dev/gin-social-network-tos
  title: Gin Social Network Service
  version: "1.0"
paths:
...
  /users/{user_id}:
    get:
      consumes:
      - application/json
      description: get user information
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserDetailInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: get user information
      tags:
      - users
  /users/edit:
    put:
      consumes:
      - application/json
      description: edit user information
      parameters:
      - description: Edit user information parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.EditUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: edit user information
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: check user's username and password
      parameters:
      - description: Login parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: Check user's username and password
      tags:
      - users
  /users/signup:
    post:
      consumes:
      - application/json
      description: create new user account
      parameters:
      - description: Create user parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: create new user account
      tags:
      - users
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      admin: "\t\t\t\t\t\t\tGrants read and write access to administrative information"
      read: "\t\t\t\t\t\t\t\tGrants read access"
      write: "\t\t\t\t\t\t\tGrants write access"
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"